chore: Update Grunt to copy dependencies into src directory and remove bower_components from Git.
I thought of keeping bower_components in Git because these dependencies are required for my app to run. However, I was able to use Grunt to copy just the files required for my app to run, so I was able to eliminate the need for bower_components inside Git.
(Note: As an additional cleanup, added load-grunt-tasks as required package so that calls to grunt.loadNpmTasks are no longer required in gruntfile.)